# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Neomed Api",
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/support/contact": {
            "post": {
                "tags": [
                    "patientApp"
                ],
                "description": "contact support with messages",
                "operationId": "contact support with messages",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "subject_id": {
                                        "type": "integer",
                                        "example": [
                                            1,
                                            2
                                        ]
                                    },
                                    "user_type": {
                                        "type": "string",
                                        "example": [
                                            "patient",
                                            "doctor"
                                        ]
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "the request to the support sent successfully , wait for response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/support/subject-types": {
            "get": {
                "tags": [
                    "patientApp"
                ],
                "description": "get support subject types",
                "operationId": "get support subject types",
                "responses": {
                    "200": {
                        "description": "support subject types fetched successfuly",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}